Migration fix

+++ database/migrations/2023_04_07_115752_create_editions_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('editions', function (Blueprint $table) {
             $table->id();
-            $table->tinyInteger('year')->unsigned();
+            $table->smallInteger('year')->unsigned();
             $table->string('city');
             $table->string('logo')->nullable();
             $table->timestamps();
